Support per-branch notification configurations
This change adds the support for multiple per-branch configurations for
a project or a specific build task. For example, users can setup a
config to notify on the staging branch, and another config to notify on
the master (default) branch to a different channel on Slack.

Also, it fixes a potential bug that when the default branch is not
master, branch matching could be incorrect, by falling back to the magic
value `<default>` from TeamCity, when filterBranchName or
branchDisplayName does not exist.